chore: 修改 dde-file-manager-plugins、dde-file-thumbnail-tool、deepin-anything-server-plugins模块单元测试目录
修改 dde-file-manager-plugins、dde-file-thumbnail-tool、deepin-anything-server-plugins模块单元测试目录

Log:


Former-commit-id: 87affc31da1810d9889831bf6f7aab1955d5299b

+++ src/dde-file-thumbnail-tool/video/funcwrapper.cpp
@@ -60,7 +60,7 @@ char *toBase64(const unsigned char *data, int size, int options)
     }
 
     if (options & OmitTrailingEquals) {
-        int new_size = out - tmp;
+        int new_size = static_cast<int>(out - tmp);
         char *new_tmp = new char[static_cast<unsigned int>(new_size) + 1];
 
         memcpy(new_tmp, tmp, static_cast<unsigned int>(new_size));
